added component for campaign cards in multiple pages
